Added dune render logic in widget and modified widget classname
